+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 **If you need help updating/installing or have found any bugs, please join [our Discord server](https://discord.gg/eGnrPqEH7U) or open a [GitHub issue here](https://github.com/Dev-CasperTheGhost/snaily-cadv3/issues/new/choose)**
 
+## 1.4.6
+
+- New: [Docker support](https://github.com/Dev-CasperTheGhost/snaily-cadv3/wiki/Installation-Guide#using-docker) - for advanced users.
+- New: Only notify an active officer to a call when they're not already assigned to a call.
+- Minor: Bug fixes with the live-map
+
 ## 1.4.5
 
 - Updated translations